Create variant.h
makerdiary nRF52832-MDK variant.h

This is basically a copy of the variant.h file for the Generic nRF52832 variant, with pins assigned for LED (RGB), button, and (important!) serial TX and RX.  I picked pins for SPI and IIC that seemed to make some sense - YMMV.  These worked well for me.  Everything else in the Generic nRF52832 is pretty much untouched and works fine with this board.