Update boards.txt
makerdiary nRF52832-MDK variant added.

Added the Makerdiary nRF52832-MDK variant information at the end of the list of nRF52832 variants.  It's basically a copy of the Generic nRF52832 variant, but points to a new folder with the modified variant.h and related files.  The target core is the Nordic Semiconductor nRF52832 SoC, so most of the information is the same.  This board has the 32kHz crystal for low power oscillator, so that option is maintained in the clock menu, and it uses the same S132 softdevice.